sadfds

+++ schemas/schemas.go
@@ -2,7 +2,7 @@ package schemas
 
 type User struct {
 	Username string `json:"username"`
-	Age      int64  `json:"age"`
+	Age      int    `json:"age"`
 	Email    string `json:"email"`
 	Role     bool   `json:"role"`
 }
